roachtest: remove version upgrade privileges roachtest
The test was created for upgrading from 21.1 to 21.2.
It is no longer necessary.

Release note: None
