+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -94,7 +94,6 @@ go_library(
         "pgx_blocklist.go",
         "pop.go",
         "predecessor_version.go",
-        "privilege_version_upgrade.go",
         "psycopg.go",
         "psycopg_blocklist.go",
         "python_helpers.go",
@@ -178,7 +177,6 @@ go_library(
         "//pkg/server/serverpb",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
-        "//pkg/sql/privilege",
         "//pkg/testutils",
         "//pkg/testutils/skip",
         "//pkg/testutils/sqlutils",
